Import of readings from all sensors works ok.

+++ projects/lone_cabbage_2017/data_management/oyster_create_insert.py
@@ -1,5 +1,8 @@
 '''
-python 3.6 code to create the lone cabbage oyster project database tables
+python 3.6 code to create the lone cabbage oyster project 'main'
+database tables to hold project, sensors, locations, info that will not
+often change, and is easily initialized via hard coding to
+facilitate set up of other applications.
 as of January 2018 or so...
 '''
 
@@ -14,8 +17,8 @@ def register_modules():
         # assume rvp office pc running windows
         modules_root="C:\\rvp\\"
     sys.path.append('{}git/citrus/modules'.format(modules_root))
-    return
-register_modules()
+    return platform_name
+platform_name = register_modules()
 import etl
 
 print("Using sys.path={}".format(repr(sys.path)))
@@ -38,6 +41,7 @@ from sqlalchemy.schema import CreateTable
 
 import sqlalchemy.sql.sqltypes
 from sqlalchemy.dialects.postgresql import ARRAY
+from sqlalchemy_tools.core.utils import drop_if_exists
 
 #
 from pathlib import Path
@@ -45,16 +49,27 @@ from pathlib import Path
 '''
 As of sqlalchemy 1.2.2, scanning google searches, it seems this is the best if
 not only apparent way I could find to check if a table exists and then to drop
-that table generically for engins psql and MySQL and possible others engines.
+that table generically for engines psql and MySQL and possible others engines.
 '''
-def drop_if_exists(engine=None,table_name=None):
-    # Drop table if it exists
+def drop_if_existsw(engine=None,table_name=None):
+
+    required_params = ['engine','table_name']
+    if not all(required_params):
+        msg=("Did not get all required params '{}'".format(required_params))
+        raise ValueError(msg)
+        # Drop table if it exists
     if engine.dialect.has_table(engine, table_name):
         conn = engine.raw_connection()
         cursor = conn.cursor()
         command = "drop table if exists {};".format(table_name)
-        cursor.execute(command)
-        conn.commit()
+        try:
+            cursor.execute(command)
+            conn.commit()
+        except Exception as ex:
+            msg = ("Table_name={}, drop statement exception='{}'"
+                .format(table_name,ex))
+            raise ValueError(msg)
+
         cursor.close()
 
 '''
@@ -212,22 +227,40 @@ def table_sensor_create(metadata=None):
     return table_object
 #def table_sensor_create
 
-def table_sensor_populate(engine=None,table_object=None):
+def table_sensor_populate(engine=None,verbosity=1):
+    sa_metadata = MetaData()
+    table_object = Table('sensor',
+            sa_metadata, autoload=True, autoload_with=engine)
 
-    l_rows = [
-        { 'project_id':1, 'location_id':1 },
-        { 'project_id':1, 'location_id':3 },
-        { 'project_id':1, 'location_id':5 },
-        { 'project_id':1, 'location_id':7 },
-        { 'project_id':1, 'location_id':9 },
-        { 'project_id':1, 'location_id':2 },
-        { 'project_id':1, 'location_id':3 },
-        { 'project_id':1, 'location_id':6 },
-        { 'project_id':1, 'location_id':8 },
-    ]
+    d_sensor_serial_manufacturer = {
+    1 : ['..02-V5602  317.', 'diver'],
+    3 : ['..00-V6916  317.', 'diver'],
+    2 : ['8814', 'star'],
+    4 : ['9058', 'star'],
+    5 : ['9060', 'star'],
+    6 : ['9061', 'star'],
+    7 : ['9035', 'star'],
+    8 : ['9062', 'star'],
+    9 : ['9036', 'star'],
+    }
+    l_rows=[]
+    for i in range (1,10):
+        d_row = {}
+        l_rows.append(d_row)
+        d_row['project_id'] = 1
+        d_row['sensor_id'] = i
+        d_row['location_id'] = i
+        d_row['observation_unit'] = 'minute'
+        d_row['observation_count'] = 60
+        d_row['serial_number'] = d_sensor_serial_manufacturer[i][0]
+        d_row['manufacturer'] = d_sensor_serial_manufacturer[i][1]
 
     #insert the l_rows
+    row_count = 0
     for row in l_rows:
+        row_count += 1
+        if verbosity > 0:
+            print("Inserting row {}= {}".format(row_count,row))
         engine.execute(table_object.insert(), row)
 
     return
@@ -259,11 +292,20 @@ def table_sensor_history_create(metadata=None):
 #def table_sensor_create
 
 def table_sensor_history_populate(engine=None,table_object=None):
+    sa_metadata = MetaData()
+    table_object = Table('sensor_history',
+            sa_metadata, autoload=True, autoload_with=engine)
 
     l_rows = [
-        { 'sensor_id':1, 'location_id':1 },
-        { 'sensor_id':2, 'location_id':9 },
-        { 'sensor_id':7, 'location_id':8 },
+        { 'sensor_id':1, 'location_id':1, 'event_type': 'placement' },
+        { 'sensor_id':2, 'location_id':2, 'event_type': 'placement'  },
+        { 'sensor_id':3, 'location_id':3, 'event_type': 'placement'  },
+        { 'sensor_id':4, 'location_id':4, 'event_type': 'placement'  },
+        { 'sensor_id':5, 'location_id':5, 'event_type': 'placement'  },
+        { 'sensor_id':6, 'location_id':6, 'event_type': 'placement'  },
+        { 'sensor_id':7, 'location_id':7, 'event_type': 'placement'  },
+        { 'sensor_id':8, 'location_id':8, 'event_type': 'placement'  },
+        { 'sensor_id':9, 'location_id':9, 'event_type': 'placement'  },
     ]
 
     #insert the l_rows
@@ -271,22 +313,45 @@ def table_sensor_history_populate(engine=None,table_object=None):
         engine.execute(table_object.insert(), row)
 
     return
-#end def table_sensor_populate
 
-def table_water_observation_populate(engine=None,table_object=None):
+'''
+Table sensor_observation is a narrower table version than table water observation.
+Table sensor_observation is updated by fixed sensors types of diver and star.
+'''
 
-    l_rows = [
-        { },
-        { },
-        { },
-    ]
+def table_sensor_observation_create(metadata=None):
+    table_name = 'sensor_observation'
+    me = '{}_create'.format(table_name)
 
-    #insert the l_rows
-    #for row in l_rows:
-    #    engine.execute(table_object.insert(), row)
+    # NOTE: we break from convention and use observation_id
+    table_object =  Table(table_name, metadata,
+      Column('{}_id'.format(table_name), Integer,
+          # NOTE do NOT use Sequence here for mysql?
+          #Sequence('{}_id_seq'.format(table_name), metadata=metadata),
+          primary_key=True, autoincrement=True,
+          comment='Automatically incremented row id.'),
+      UniqueConstraint('{}_id'.format(table_name),
+          name='uq1_{}'.format(table_name) ),
+      Column('sensor_id', Integer),
+      Column('observation_datetime', DateTime),
+      Column('location_id', Integer, default=1),
+      UniqueConstraint('sensor_id','observation_datetime',
+          name='uq2_{}'.format(table_name) ),
+      # location_id can be derived, maybe no need to populate via imports?
+      Column('conductivity_mS_cm', Float),
+      Column('pressure_cm', Float),
+      Column('salinity_psu', Float),
+      Column('temperature_c', Float),
+      Column('sound_velocity_m_sec', Float),
+      ForeignKeyConstraint(
+        ['sensor_id'], ['sensor.sensor_id'],
+        name='fk_{}_sensor_id'.format(table_name)),
+      ForeignKeyConstraint(
+        ['location_id'], ['location.location_id'],
+        name='fk_{}_location_id'.format(table_name)),
+      )
 
-    return
-#end def table_sensor_populate
+    return table_object
 
 def table_water_observation_create(metadata=None):
     table_name = 'water_observation'
@@ -315,9 +380,12 @@ def table_water_observation_create(metadata=None):
       Column('specific_conductance_us_cm_25c', Float),
       Column('specific_conductance_ms_cm_25c', Float),
       Column('salinity_g_kg', Float),
+      Column('salinity_psu', Float),
       Column('temperature_c', Float),
       Column('pressure_psi', Float),
-      Column('sound_velocity_m_s', Float),
+      Column('pressure_cm', Float),
+      Column('conductivity_mS_cm', Float),
+      Column('sound_velocity_m_sec', Float),
       Column('note', String(20),
              comment='Short note on observation'),
       ForeignKeyConstraint(
@@ -331,10 +399,24 @@ def table_water_observation_create(metadata=None):
     return table_object
 #end def table_water_observation_create
 
+'''
+<summary name="tables_create">
+</summary>
+
+'''
 def tables_create(engine=None, metadata=None):
     #metadata = MetaData()
     d_name_table = {}
 
+    # Drop more dependent tables first
+    drop_if_exists(engine=engine, table_name='sensor_observation')
+    drop_if_exists(engine=engine, table_name='water_observation')
+    drop_if_exists(engine=engine, table_name='sensor_history')
+    drop_if_exists(engine=engine, table_name='sensor')
+    drop_if_exists(engine=engine, table_name='location')
+    drop_if_exists(engine=engine, table_name='project')
+
+    # Create most independent tables first
     d_name_table['project'] = table_project_create(metadata=metadata)
 
     ptable = d_name_table['project']
@@ -347,6 +429,8 @@ def tables_create(engine=None, metadata=None):
     d_name_table['sensor_history'] = table_sensor_history_create(
         metadata=metadata)
     d_name_table['water_observation'] = table_water_observation_create(
+        metadata=metadata)
+    d_name_table['sensor_observation'] = table_sensor_observation_create(
         metadata=metadata)
 
     metadata.create_all(engine, checkfirst=False)
@@ -361,16 +445,11 @@ def tables_populate(engine=None, metadata=None,d_name_table=None):
         table_object=d_name_table['project'])
     table_location_populate(engine=engine,
         table_object=d_name_table['location'])
-    table_sensor_populate(engine=engine,
-        table_object=d_name_table['sensor'])
+    table_sensor_populate(engine=engine)
     table_sensor_history_populate(engine=engine,
         table_object=d_name_table['sensor_history'])
-    try:
-        t = d_name_table['water_observation']
-    except:
-        raise ValueError()
-
-    table_water_observation_populate(engine=engine, table_object=t)
+    # No need to initialize water_observtion table, as an import
+    # program is now working
 
     return
 
@@ -378,6 +457,8 @@ def tables_populate(engine=None, metadata=None,d_name_table=None):
 
 # MAIN CODE
 def run(env=None):
+    me="run (oyster_tables_create)"
+    print("STARTING: {}: starting".format(me))
     if env == 'uf':
         #something all messet up.. THIS works for UF! using env of uf...
         engine_nick_name = 'uf_local_mysql_marshal1'
@@ -396,10 +477,17 @@ def run(env=None):
     d_name_table = tables_create(engine=engine,metadata=metadata)
     tables_populate(engine=engine,metadata=metadata,d_name_table=d_name_table)
 
+    print("ENDING: {}: ending".format(me))
+    return
+#end def run
+
 #
 test = 1
+
 if test == 1 :
-    env = 'uf'
-    env = 'home'
-    env = 'uf'
+    if platform_name == 'linux':
+        env = 'home'
+    else:
+        env = 'uf'
+
     run(env=env)
